+++ assets/test/dom_test.js
@@ -120,7 +120,7 @@ describe("DOM", () => {
     })
   })
 
-  describe("findParentCIDs", () => {
+  describe("findExistingParentCIDs", () => {
     test("returns only parent cids", () => {
       let view = tag("div", {}, `
         <div data-phx-main="true"
@@ -133,19 +133,17 @@ describe("DOM", () => {
       `)
       document.body.appendChild(view)
 
-      expect(DOM.findParentCIDs(view, [1, 2, 3])).toEqual(new Set([1, 2, 3]))
-
       view.appendChild(tag("div", {"data-phx-component": 1}, `
         <div data-phx-component="2"></div>
       `))
-      expect(DOM.findParentCIDs(view, [1, 2, 3])).toEqual(new Set([1, 3]))
+      expect(DOM.findExistingParentCIDs(view, [1, 2])).toEqual(new Set([1]))
 
       view.appendChild(tag("div", {"data-phx-component": 1}, `
         <div data-phx-component="2">
           <div data-phx-component="3"></div>
         </div>
       `))
-      expect(DOM.findParentCIDs(view, [1, 2, 3])).toEqual(new Set([1]))
+      expect(DOM.findExistingParentCIDs(view, [1, 2, 3])).toEqual(new Set([1]))
     })
   })