+++ config/e2e.exs
@@ -1,0 +1,3 @@
+import Config
+
+config :logger, :level, :error