+++ config/test.exs
@@ -1,0 +1,6 @@
+import Config
+
+config :logger, :level, :debug
+config :logger, :backends, []
+
+config :phoenix_live_view, enable_expensive_runtime_checks: true