+++ lib/phoenix_live_view/async.ex
@@ -3,44 +3,73 @@ defmodule Phoenix.LiveView.Async do
 
   alias Phoenix.LiveView.{AsyncResult, Socket, Channel}
 
-  defp validate_function_env(func, op, env) do
-    warn =
-      # TODO: Remove conditional once we require Elixir v1.14+
-      if Version.match?(System.version(), ">= 1.14.0") do
-        fn msg, env -> IO.warn(msg, env) end
-      else
-        fn msg, _env -> IO.warn(msg) end
-      end
+  defp warn_socket_access(op, warn) do
+    warn.("""
+    you are accessing the LiveView Socket inside a function given to #{op}.
 
+    This is an expensive operation because the whole socket is copied to the new process.
+
+    Instead of:
+
+        #{op}(socket, :key, fn ->
+          do_something(socket.assigns.my_assign)
+        end)
+
+    You should do:
+
+        my_assign = socket.assigns.my_assign
+
+        #{op}(socket, :key, fn ->
+          do_something(my_assign)
+        end)
+
+    For more information, see https://hexdocs.pm/elixir/1.16.1/process-anti-patterns.html#sending-unnecessary-data.
+    """)
+  end
+
+  # this is not private to prevent the unused function warning as we only
+  # call this function when enable_expensive_runtime_checks is set
+  def warn_assigns_access(op, warn) do
+    warn.("""
+    you are accessing an assigns map inside a function given to #{op}.
+
+    This is an expensive operation because the whole map is copied to the new process.
+
+    Instead of:
+
+        #{op}(socket, :key, fn ->
+          do_something(assigns.my_assign)
+        end)
+
+    You should do:
+
+        my_assign = assigns.my_assign
+
+        #{op}(socket, :key, fn ->
+          do_something(my_assign)
+        end)
+
+    For more information, see https://hexdocs.pm/elixir/1.16.1/process-anti-patterns.html#sending-unnecessary-data.
+    """)
+  end
+
+  defp validate_function_env(func, op, env) do
     # prevent false positives, for example
     # start_async(socket, :foo, function_that_returns_the_anonymous_function(socket))
     if match?({:&, _, _}, func) or match?({:fn, _, _}, func) do
       Macro.prewalk(func, fn
-        {:socket, meta, nil} ->
-          warn.(
-            """
-            you are accessing the LiveView Socket inside a function given to #{op}.
+        {:socket, meta, _} ->
+          warn_socket_access(op, fn msg ->
+            # TODO: Remove conditional once we require Elixir v1.14+
+            meta =
+              if Version.match?(System.version(), ">= 1.14.0") do
+                Keyword.take(meta, [:line, :column]) ++ [line: env.line, file: env.file]
+              else
+                Macro.Env.stacktrace(env)
+              end
 
-            This is an expensive operation because the whole socket is copied to the new process.
-
-            Instead of:
-
-                #{op}(socket, :key, fn ->
-                  do_something(socket.assigns.my_assign)
-                end)
-
-            You should do:
-
-                my_assign = socket.assigns.my_assign
-
-                #{op}(socket, :key, fn ->
-                  do_something(my_assign)
-                end)
-
-            For more information, see https://hexdocs.pm/elixir/1.16.1/process-anti-patterns.html#sending-unnecessary-data.
-            """,
-            Keyword.take(meta, [:line, :column]) ++ [line: env.line, file: env.file]
-          )
+            IO.warn(msg, meta)
+          end)
 
         other ->
           other
@@ -50,18 +79,45 @@ defmodule Phoenix.LiveView.Async do
     :ok
   end
 
+  if Application.compile_env(:phoenix_live_view, :enable_expensive_runtime_checks, false) do
+    defp validate_function_env(func, op) do
+      {:env, variables} = Function.info(func, :env)
+
+      cond do
+        Enum.any?(variables, &match?(%Phoenix.LiveView.Socket{}, &1)) ->
+          warn_socket_access(op, fn msg -> IO.warn(msg) end)
+
+        Enum.any?(variables, &match?(%{__changed__: _}, &1)) ->
+          warn_assigns_access(op, fn msg -> IO.warn(msg) end)
+
+        true ->
+          :ok
+      end
+    end
+  else
+    defp validate_function_env(_func, _op), do: :ok
+  end
+
   def start_async(socket, key, func, opts, env) do
     validate_function_env(func, :start_async, env)
 
     quote do
-      Phoenix.LiveView.Async.run_async_task(
+      Phoenix.LiveView.Async.start_async(
         unquote(socket),
         unquote(key),
         unquote(func),
-        :start,
         unquote(opts)
       )
     end
+  end
+
+  def start_async(%Socket{} = socket, key, func, opts) when is_function(func, 0) do
+    # runtime check
+    if Phoenix.LiveView.connected?(socket) do
+      validate_function_env(func, :start_async)
+    end
+
+    run_async_task(socket, key, func, :start, opts)
   end
 
   def assign_async(socket, key_or_keys, func, opts, env) do
@@ -80,6 +136,11 @@ defmodule Phoenix.LiveView.Async do
   def assign_async(%Socket{} = socket, key_or_keys, func, opts)
       when (is_atom(key_or_keys) or is_list(key_or_keys)) and
              is_function(func, 0) do
+    # runtime check
+    if Phoenix.LiveView.connected?(socket) do
+      validate_function_env(func, :assign_async)
+    end
+
     keys = List.wrap(key_or_keys)
 
     # verifies result inside task