+++ lib/phoenix_live_view/html_algebra.ex
@@ -3,9 +3,6 @@ defmodule Phoenix.LiveView.HTMLAlgebra do
 
   import Inspect.Algebra, except: [format: 2]
 
-  # TODO: Remove it after versions before Elixir 1.13 are no longer supported.
-  @compile {:no_warn_undefined, Code}
-
   @languages ~w(style script)
 
   # The formatter has two modes: