+++ priv/static/phoenix_live_view.cjs.js
@@ -700,14 +700,17 @@ var DOM = {
   findPhxChildren(el, parentId) {
     return this.all(el, `${PHX_VIEW_SELECTOR}[${PHX_PARENT_ID}="${parentId}"]`);
   },
-  findParentCIDs(node, cids) {
-    let initial = new Set(cids);
-    let parentCids = cids.reduce((acc, cid) => {
-      let selector = `[${PHX_COMPONENT}="${cid}"] [${PHX_COMPONENT}]`;
-      this.filterWithinSameLiveView(this.all(node, selector), node).map((el) => parseInt(el.getAttribute(PHX_COMPONENT))).forEach((childCID) => acc.delete(childCID));
-      return acc;
-    }, initial);
-    return parentCids.size === 0 ? new Set(cids) : parentCids;
+  findExistingParentCIDs(node, cids) {
+    let parentCids = new Set();
+    let childrenCids = new Set();
+    cids.forEach((cid) => {
+      this.filterWithinSameLiveView(this.all(node, `[${PHX_COMPONENT}="${cid}"]`), node).forEach((parent) => {
+        parentCids.add(cid);
+        this.all(parent, `[${PHX_COMPONENT}]`).map((el) => parseInt(el.getAttribute(PHX_COMPONENT))).forEach((childCID) => childrenCids.add(childCID));
+      });
+    });
+    childrenCids.forEach((childCid) => parentCids.delete(childCid));
+    return parentCids;
   },
   filterWithinSameLiveView(nodes, parent) {
     if (parent.querySelector(PHX_VIEW_SELECTOR)) {
@@ -3329,7 +3332,7 @@ var View = class {
     let phxChildrenAdded = false;
     if (this.rendered.isComponentOnlyDiff(diff)) {
       this.liveSocket.time("component patch complete", () => {
-        let parentCids = dom_default.findParentCIDs(this.el, this.rendered.componentCIDs(diff));
+        let parentCids = dom_default.findExistingParentCIDs(this.el, this.rendered.componentCIDs(diff));
         parentCids.forEach((parentCID) => {
           if (this.componentPatch(this.rendered.getComponent(diff, parentCID), parentCID)) {
             phxChildrenAdded = true;