+++ test/phoenix_live_view/integrations/assign_async_test.exs
@@ -51,8 +51,14 @@ defmodule Phoenix.LiveView.AssignAsyncTest do
     end
 
     test "lv exit brings down asyncs", %{conn: conn} do
+      Process.register(self(), :assign_async_test_process)
       {:ok, lv, _html} = live(conn, "/assign_async?test=lv_exit")
       lv_ref = Process.monitor(lv.pid)
+
+      receive do
+        :async_ready -> :ok
+      end
+
       async_ref = Process.monitor(Process.whereis(:lv_exit))
       send(lv.pid, :boom)
 
@@ -61,7 +67,13 @@ defmodule Phoenix.LiveView.AssignAsyncTest do
     end
 
     test "cancel_async", %{conn: conn} do
+      Process.register(self(), :assign_async_test_process)
       {:ok, lv, _html} = live(conn, "/assign_async?test=cancel")
+
+      receive do
+        :async_ready -> :ok
+      end
+
       async_ref = Process.monitor(Process.whereis(:cancel))
       send(lv.pid, :cancel)
 
@@ -153,8 +165,14 @@ defmodule Phoenix.LiveView.AssignAsyncTest do
     end
 
     test "lv exit brings down asyncs", %{conn: conn} do
+      Process.register(self(), :assign_async_test_process)
       {:ok, lv, _html} = live(conn, "/assign_async?test=lc_lv_exit")
       lv_ref = Process.monitor(lv.pid)
+
+      receive do
+        :async_ready -> :ok
+      end
+
       async_ref = Process.monitor(Process.whereis(:lc_exit))
       send(lv.pid, :boom)
 
@@ -163,7 +181,13 @@ defmodule Phoenix.LiveView.AssignAsyncTest do
     end
 
     test "cancel_async", %{conn: conn} do
+      Process.register(self(), :assign_async_test_process)
       {:ok, lv, _html} = live(conn, "/assign_async?test=lc_cancel")
+
+      receive do
+        :async_ready -> :ok
+      end
+
       async_ref = Process.monitor(Process.whereis(:lc_cancel))
 
       Phoenix.LiveView.send_update(lv.pid, Phoenix.LiveViewTest.AssignAsyncLive.LC,