+++ test/phoenix_live_view/integrations/connect_test.exs
@@ -1,5 +1,5 @@
 defmodule Phoenix.LiveView.ConnectTest do
-  use ExUnit.Case, async: false
+  use ExUnit.Case, async: true
   import Phoenix.LiveViewTest
   import Phoenix.ConnTest