+++ test/phoenix_live_view/integrations/hooks_test.exs
@@ -1,5 +1,5 @@
 defmodule Phoenix.LiveView.HooksTest do
-  use ExUnit.Case, async: false
+  use ExUnit.Case, async: true
 
   import Phoenix.ConnTest
   import Phoenix.LiveViewTest
@@ -295,8 +295,30 @@ defmodule Phoenix.LiveView.HooksTest do
     assert render_async(lv) =~ "task:o.\n"
   end
 
+  test "attach/detach_hook with a handle_event live component socket", %{conn: conn} do
+    {:ok, lv, _html} = live(conn, "/lifecycle/components/handle_event")
+    lv |> element("#attach") |> render_click()
+    lv |> element("#hook") |> render_click()
+    assert render_async(lv) =~ "counter: 1"
+
+    lv |> element("#hook") |> render_click()
+    assert render_async(lv) =~ "counter: 2"
+
+    lv |> element("#detach-component-hook") |> render_click()
+    Process.flag(:trap_exit, true)
+
+    assert ExUnit.CaptureLog.capture_log(fn ->
+             try do
+               lv |> element("#hook") |> render_click()
+             catch
+               :exit, _ -> :ok
+             end
+           end) =~
+             "** (UndefinedFunctionError) function Phoenix.LiveViewTest.HooksEventComponent.handle_event/3 is undefined"
+  end
+
   test "attach_hook raises when given a live component socket", %{conn: conn} do
-    {:ok, lv, _html} = live(conn, "/lifecycle/components")
+    {:ok, lv, _html} = live(conn, "/lifecycle/components/handle_info")
 
     assert HooksLive.exits_with(lv, ArgumentError, fn ->
              lv |> element("#attach") |> render_click()
@@ -304,7 +326,7 @@ defmodule Phoenix.LiveView.HooksTest do
   end
 
   test "detach_hook raises when given a live component socket", %{conn: conn} do
-    {:ok, lv, _html} = live(conn, "/lifecycle/components")
+    {:ok, lv, _html} = live(conn, "/lifecycle/components/handle_info")
 
     assert HooksLive.exits_with(lv, ArgumentError, fn ->
              lv |> element("#detach") |> render_click()